sailor-agent/requirements.txt
# anydata_sdk==0.0.2.694591
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.0.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.3.2
dataclasses-json==0.6.7
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx>=0.27.0
idna==3.7
isort==5.13.2
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.77
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.68.2
opensearch-py==2.4.2
orjson==3.10.4
packaging==23.2
pandas==2.2.2
platformdirs==4.2.2
pyahocorasick==2.0.0
pycryptodome==3.20.0
pydantic==2.7.3
pydantic-settings==2.1.0
pydantic_core==2.18.4
pylint==3.0.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
regex==2024.5.15
requests==2.32.3
scikit-learn==1.3.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sse-starlette==1.8.2
starlette>=0.37.2,<0.39.0
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.23.2
xmlrunner==1.7.7
yarl==1.9.4
altair==5.4.0
commentjson==0.9.0
plotly==5.23.0
jsonref==1.1.0
vegafusion==1.6.9
func_timeout==4.3.5

kweaver-data-retrieval==0.1.2
# af_agent==0.2.0a2
sqlparse==0.5.1
sql_metadata==2.13.0
snowflake-id
dmPython==2.5.8
dmSQLAlchemy==2.0.11
sseclient-py==1.8.0
circuitbreaker